+++ lib/plugins/aws/package/compile/events/alb/index.js
@@ -4,7 +4,7 @@ const BbPromise = require('bluebird');
 
 const validate = require('./lib/validate');
 const compileTargetGroups = require('./lib/target-groups');
-const compileListenerRules = require('./lib/listenerRules');
+const compileListenerRules = require('./lib/listener-rules');
 const compilePermissions = require('./lib/permissions');
 
 function defineArray(schema) {
